+++ src/fastsqrt/fastsqrt.c
@@ -11,10 +11,11 @@
  * NO WARRANTY!
  */
 
+#include <inttypes.h>
 #include <math.h>
 #include "fastsqrt.h"
 
-static long sqrt_table[0x100];
+static int32_t sqrt_table[0x100];
 
 void build_sqrt_table(void) {
     unsigned short i;
