+++ src/object.c
@@ -106,7 +106,7 @@ struct object {
      * the game time at which that was sent. */
     /* [pva == position, velocity, acceleration] */
     struct {
-	unsigned long time;
+	uint32_t time;
 	float x, y;
 	float xv, yv;
 	float xa, ya;
@@ -1453,7 +1453,7 @@ void object_do_physics (object_t *obj, map_t *map)
 
 
 void object_set_auth_info (object_t *obj,
-			   unsigned long time,
+			   uint32_t time,
 			   float x, float y,
 			   float xv, float yv,
 			   float xa, float ya)
@@ -1468,7 +1468,7 @@ void object_set_auth_info (object_t *obj,
 }
 
 
-void object_do_simulation (object_t *obj, unsigned long curr_time)
+void object_do_simulation (object_t *obj, uint32_t curr_time)
 {
     int interval;
